requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
Requests==2.32.3
selenium==4.25.0
transformers==4.46.0
torch==2.5.0
webdriver-manager==4.0.2
pillow==11.0.0
opencv-python==4.10.0.84
scikit-image==0.24.0
google-generativeai==0.8.3
python-telegram-bot==21.7
azure-keyvault-secrets==4.10.0
google-cloud-secret-manager==2.24.0
boto3==1.40.41